feat: Rename example timeline code as "honeycomb"
